feat: Add-Middleware Login and Logout

+++ routes/web.php
@@ -1,7 +1,7 @@
 <?php
 
 use App\Http\Controllers\HomeController;
-use App\Http\Controllers\LoginController;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\PostController;
 use Illuminate\Support\Facades\Route;
 
@@ -19,15 +19,17 @@ use Illuminate\Support\Facades\Route;
 Route::get("/home", [HomeController::class, "index"])->name("home");
 
 // Login Auth
-Route::get("/", [LoginController::class, "index"])->name("auth.login");
+Route::get("/", [AuthController::class, "login"])->name("auth.login")->middleware("guest");
+Route::post("/login", [AuthController::class, "authenticating"])->name("authenticating")->middleware("guest");
+Route::get("/logout", [AuthController::class, "logout"])->name("logout")->middleware("auth");
 
 // Dashboard Admin
 // Route::prefix("admin")->group(function (){
-    Route::get("/news", [PostController::class, "index"])->name("admin.index");
-    Route::get("/news/{id}", [PostController::class, "show"])->name("post.detail");
-    Route::get("/news-add", [PostController::class, "create"])->name("admin.create");
-    Route::post("/new" ,[PostController::class, "store"])->name("store-data");
-    Route::get("/news-edit/{id}", [PostController::class, "edit"])->name("admin.edit");
-    Route::put("/new/{id}", [PostController::class, "update"])->name("update-data");
-    Route::delete('/news-delete/{id}', [PostController::class, "destroy"])->name("delete-data");
+    Route::get("/news", [PostController::class, "index"])->name("admin.index")->middleware("auth");
+    Route::get("/news/{id}", [PostController::class, "show"])->name("post.detail")->middleware("auth");
+    Route::get("/news-add", [PostController::class, "create"])->name("admin.create")->middleware("auth");
+    Route::post("/new" ,[PostController::class, "store"])->name("store-data")->middleware("auth");
+    Route::get("/news-edit/{id}", [PostController::class, "edit"])->name("admin.edit")->middleware("auth");
+    Route::put("/new/{id}", [PostController::class, "update"])->name("update-data")->middleware("auth");
+    Route::delete('/news-delete/{id}', [PostController::class, "destroy"])->name("delete-data")->middleware("auth");
 // });
